+++ src/spotmicroai/diagnostics/diagnostics.py
@@ -26,8 +26,8 @@ from spotmicroai.constants import (
 from spotmicroai.servo import JointType
 from spotmicroai.servo._servo import Servo
 from spotmicroai.servo._servo_factory import ServoFactory
-from spotmicroai.setup_app import theme as THEME, ui_utils
-import spotmicroai.setup_app.labels as LABELS
+from spotmicroai.ui import theme as THEME, ui_utils
+import spotmicroai.labels as LABELS
 
 
 class DiagnosticsWizard:
@@ -154,7 +154,7 @@ class DiagnosticsWizard:
         except curses.error:
             pass
 
-    def sweep_servos(self, servo_names: List[ServoName], start_angle: int, end_angle: int) -> bool:
+    def sweep_servos(self, servo_names: List[ServoName], start_angle: float, end_angle: float) -> bool:
         """Sweep a group of servos from start_angle to end_angle.
 
         Args:
